extract buttons in their own component
