+++ src/main/webapp/app/course/dashboards/assessment-dashboard/assessment-dashboard.component.ts
@@ -22,13 +22,6 @@ import { QuizExercise } from 'app/entities/quiz/quiz-exercise.model';
 import { AssessmentDashboardInformationEntry } from './assessment-dashboard-information.component';
 import { TutorIssue, TutorIssueComplaintsChecker, TutorIssueRatingChecker, TutorIssueScoreChecker } from 'app/course/dashboards/assessment-dashboard/tutor-issue';
 import { TutorLeaderboardElement } from 'app/shared/dashboards/tutor-leaderboard/tutor-leaderboard.model';
-import { onError } from 'app/shared/util/global.utils';
-import { ArtemisTranslatePipe } from 'app/shared/pipes/artemis-translate.pipe';
-import { HttpErrorResponse } from '@angular/common/http';
-import dayjs from 'dayjs';
-import { StudentExamService } from 'app/exam/manage/student-exams/student-exam.service';
-import { StudentExam } from 'app/entities/student-exam.model';
-import { tap } from 'rxjs/operators';
 
 @Component({
     selector: 'jhi-courses',
@@ -77,12 +70,6 @@ export class AssessmentDashboardComponent implements OnInit {
 
     tutorIssues: TutorIssue[] = [];
 
-    isLoading: boolean;
-    isExamOver = false;
-    longestWorkingTime?: number;
-    studentExams: StudentExam[];
-    isExamStarted = false;
-
     constructor(
         private courseService: CourseManagementService,
         private exerciseService: ExerciseService,
@@ -92,15 +79,12 @@ export class AssessmentDashboardComponent implements OnInit {
         private route: ActivatedRoute,
         private guidedTourService: GuidedTourService,
         private sortService: SortService,
-        private artemisTranslatePipe: ArtemisTranslatePipe,
-        private studentExamService: StudentExamService,
     ) {}
 
     /**
      * On init set the courseID, load all exercises and statistics for tutors and set the identity for the AccountService.
      */
     ngOnInit(): void {
-        this.isLoading = true;
         this.courseId = Number(this.route.snapshot.paramMap.get('courseId'));
         this.examId = Number(this.route.snapshot.paramMap.get('examId'));
         this.isExamMode = !!this.examId;
@@ -183,17 +167,6 @@ export class AssessmentDashboardComponent implements OnInit {
                     this.computeIssuesWithTutorPerformance();
                 },
                 (response: string) => this.onError(response),
-            );
-            this.isExamStarted = this.exam.startDate ? this.exam.startDate.isBefore(dayjs()) : false;
-            this.studentExamService.findAllForExam(this.courseId, this.examId).pipe(
-                tap((res) => {
-                    this.setStudentExams(res.body);
-                    this.longestWorkingTime = Math.max.apply(
-                        null,
-                        this.studentExams.map((studentExam) => studentExam.workingTime),
-                    );
-                    this.calculateIsExamOver();
-                }),
             );
         } else {
             this.courseService.getCourseWithInterestingExercisesForTutors(this.courseId).subscribe(
@@ -419,73 +392,5 @@ export class AssessmentDashboardComponent implements OnInit {
 
     asQuizExercise(exercise: Exercise): QuizExercise {
         return exercise as QuizExercise;
-    }
-
-    /**
-     * Evaluates all the quiz exercises that belong to the exam
-     */
-    evaluateQuizExercises() {
-        this.isLoading = true;
-        this.examManagementService.evaluateQuizExercises(this.courseId, this.examId).subscribe(
-            (res) => {
-                this.alertService.success('artemisApp.studentExams.evaluateQuizExerciseSuccess', { number: res?.body });
-                this.isLoading = false;
-            },
-            (err: HttpErrorResponse) => {
-                this.handleError('artemisApp.studentExams.evaluateQuizExerciseFailure', err);
-                this.isLoading = false;
-            },
-        );
-    }
-
-    assessUnsubmittedExamModelingAndTextParticipations() {
-        this.isLoading = true;
-        this.examManagementService.assessUnsubmittedExamModelingAndTextParticipations(this.courseId, this.examId).subscribe(
-            (res) => {
-                this.alertService.success('artemisApp.studentExams.assessUnsubmittedStudentExamsSuccess', { number: res?.body });
-                this.isLoading = false;
-            },
-            (err: HttpErrorResponse) => {
-                this.handleError('artemisApp.studentExams.assessUnsubmittedStudentExamsFailure', err);
-                this.isLoading = false;
-            },
-        );
-    }
-
-    /**
-     * Shows the translated error message if an error key is available in the error response. Otherwise it defaults to the generic alert.
-     * @param translationString the string identifier in the translation service for the text. This is ignored if the response does not contain an error message or error key.
-     * @param err the error response
-     */
-    private handleError(translationString: string, err: HttpErrorResponse) {
-        let errorDetail;
-        if (err?.error && err.error.errorKey) {
-            errorDetail = this.artemisTranslatePipe.transform(err.error.errorKey);
-        } else {
-            errorDetail = err?.error?.message;
-        }
-        if (errorDetail) {
-            this.alertService.error(translationString, { message: errorDetail });
-        } else {
-            // Sometimes the response does not have an error field, so we default to generic error handling
-            onError(this.alertService, err);
-        }
-    }
-
-    calculateIsExamOver() {
-        if (this.longestWorkingTime && this.exam) {
-            const startDate = dayjs(this.exam.startDate);
-            let endDate = startDate.add(this.longestWorkingTime, 'seconds');
-            if (this.exam.gracePeriod) {
-                endDate = endDate.add(this.exam.gracePeriod!, 'seconds');
-            }
-            this.isExamOver = endDate.isBefore(dayjs());
-        }
-    }
-
-    private setStudentExams(studentExams: any): void {
-        if (studentExams) {
-            this.studentExams = studentExams;
-        }
     }
 }
